requirements.txt
python-dotenv
psycopg2-binary
openai
flask
sqlalchemy